requirements.txt
Django==2.2.4
djangorestframework==3.10.2
django-oauth-toolkit==1.2.0
django-filter==2.2.0
mock==3.0.5
Pillow==6.1.0